+++ dao-contracts/tests/contracts/mod.rs
@@ -1,0 +1,3 @@
+mod test_repo_voter;
+mod test_reputation;
+mod test_variable_repository;
