+++ dao-contracts/tests/modules/mod.rs
@@ -1,0 +1,2 @@
+mod test_governance_voting;
+mod test_owner;
