+++ dao-contracts/tests/units.rs
@@ -1,0 +1,2 @@
+mod contracts;
+mod modules;
